Add Hashgraph Reset logic
We want to be able to clean the Hashgraph data and start afresh with
new Roots. The Reset method does just that. It cleans the Store and sets
the Roots to what it receives as parameter.

When we receive WireEvents from a peer and convert them to Events, we already
have wire info. Hence, we added a flag to the InsertEvent method to specify
wheter or not to set wire info.