Implement Node as a State-Machine with CatchingUp
Nodes can be in one of three states: Babbling, CatchingUp or
Shutdown. The Babbling state is the normal mode of operation
where Nodes periodicaly sync with other Nodes. Upon receiving
a SyncLimit response from a peer, they swith to the CatchingUp
state. SyncLimit indicates that the requesting Node is too far
behind; There would be too many new Events to process so we
need to fast-forward to a later Frame and discard parts of the
Hashgraph. Once the Node is done catching-up, it switches back
to the Babbling mode.