Change Shutdown logic
There was some redundancy in how we recorded a Shutdown signal. We
removed the shutdown flag and corresponding mutex. Now we only use
the Shutdown state of the Node state-machine.

We also wait for all goroutines to finish before actually shutting
down. The Shutdown methods should thus provide a way to gracefully
stop a node.

ATTENTION: A node cannot be restarted after it was shutdown because
the Transport is closed. If we need to test flicking nodes on and off
it is best to shutdown/kill and instantiate a new node.