Implement ControlTimer
Babble only gossips when there are uncommitted transactions in the
system. As long as there are uncommitted transactions, a timer triggers
gossip at bounded random intervals. When all pending transactions are
commited, the timer is stopped. It is restarted when new Events are
received from peers or upon receiving a SubmitTx from the AppProxy.
So the timer is manipulated and consumed by different goroutines.

The ControlTimer is an object that allows to coordinate the timer
from multiple goroutines.