Separate Docker files from demo
The docker folder used to contain a mix of docker and demo files.
We created a separate folder for the demo. It uses Docker images
which can be create from the docker/ folder. We also modified the
babble Dockerimage to use alpine instead of ubuntu - it is a lot
lighter.