Remove LastFrom from DB
LastFrom is something pertaining to the 'tip' of the Hashgraph. This
information belongs in the InmemStore cache. The DB should be used
to retrieve old items only.

More generaly, the DB cannot exist without the InmemStore cache. Together
they form the "tail" and the "tip" of the Hashgraph Store. As such,
even the Reset() method only affects the "tip" of the Hashgraph.