Implement topological index of Events in BadgerStore
Events are normally indexed by their Hash which is random so it was
impossible to retrieve the Events in the order they were inserted.
Our solution is to keep a separate mapping.

[topological index] => [event hash]
[event hash]        => [event bytes]

The topological index counter is a value that is specific to every
node. It is never used in any communication among nodes. It is a
value that helps local computations only.