Allow unspecified IP in node_addr option
The TCP transport used to enforce that the bind address (where babble
listens to other babble nodes) be explicitely specified. Now we make
it possible to specify something of the sort ':1337' so that the transport
will use all available interfaces.

We also updated the Dummy app and the terraform scripts because there
was a bug with launching the Dummy client asynchronously.