Revert changes to TCP transport
Do not allow unspecified bind address. We had allowed it for the
wrong reasons. The difficulty is when the node is started behind a
NAT router; the bind address might be different from the public IP
listed in peers.json. This leads to confusion but needs to be addressed
in a smarter way.