Extend FindOrder to save Blocks in Store
We extended the Store interface with methods to Get/Set Blocks.
The FindOrder method used to commit batches of Events when the
fame of all witnesses in a round was decided. Now, for all Events
with an equal RoundReceived, the method also packages their
transactions into a Block and saves it in the Store.