Change Mobile API to commit blocks
Until now the CommitHandler expected to receive transactions because
gomobile does not support exporting arrays of arrays. Now we have
chosen to serialize each Block into a single byte array (using json
encoding) and pass that on to the CommitHandler. The app is expected
to deserialize the byte array back into a Block and to extract and
process each transaction sequentially.