Implement FastForward
When a node attempts to sync from another node and receives a
SyncLimit response (indicating that there would be too many Events
to download), it goes into the "CatchingUp" state where it will
fast-forward to the latest consensensus Frame of the hashgraph.
It sends a FastForward request to another peer, which returns the
latest consensus Block (with super-majority of signatures) with
corresponding Frame. The Frame is checked against the Block's
FrameHash and used to Reset the Hashgraph to the correspondng
state. The node then resumes participating as usual.

TODO: We can fast-forward the hashgraph BUT not the application.
The next step is to implement a snapshot feature whereby Babble
can obtain a Snapshot of the application, and send it along with
the Block and Frame in response to a FastForward request, and the
requesting node can update the application state from that
snapshot.