Replace hg CommitCh with Callback.
Until now, committing a block to the application was an asynchronous
operation (with respect to the hashgraph). This was not a very good
idea, so we replaced it with a synchronous callback mechanism. The
hashgraph is now able to process the Commit response, which will be
useful in validating InternalTransactions before applying them.