Peer ID int -> uint32
Peer IDs, as computed by the common.Hash32() function, are 32 bit
unsigned integers. If we define them as int, instead of uint32
explicitely, we won't be able to support 32bit machines. Indeed,
in golang, 'int' stands for  'int64' or 'int32' depending on the
underlying architecture, and assigning a uint32 to a int32 would
break. Some old android phones are 32 bit, and this change allows
running Babble on them too.