Add node tests
These tests show that the Join problem is actually a problem with
FastSync. There is a scenario when a Reset hashgraph can start
assigning the wrong rounds to Events. This scenario cannot occur
with a cluster of 4 nodes with 3 active, that we have been using in
most FastSync tests. The Join tests pass in the 4/3 configuration
as well.