Add welcome transaction in commit callback
After an InternalTransaction is processed and a new peer is added,
we create a "welcome" transactions to force the nodes from the
previous peer-set to quickly reach a round above the new peer's
accepted-round. This is a quick fix for issue 104, where a joining
node had to wait for the other nodes to reach its accepted-round
before pitching into the gossip.