Change accepted-round from R+4 to R+6
Resolves issue #107. InternalTransactions should affect the peer-set
of rounds R+6 and above, where R is the round-received of the event
containing the InternalTransaction. The previous value of R+4 was
theoretically wrong, not taking into account hashgraphs that do not
have a witness at R+3 (important condition of lemma 5.17). The
dynamic-participants section in docs explains the new value.