Force AnchorBlock for AcceptedRound
When a new peer joins the group, it needs to fast-forward above its
accepted-round. So we need an AnchorBlock above the accepted-round.
In this commit we added a link between the core object and the
underlying hashgraph to force the saving and committing blocks that
are required as anchors, even if they contain no transactions.