Optional Bootstrap
Until now, when a badgerdb already existed, the node would
systematically load all the data (aka bootstrap). Now we made this
optional. If the --bootstrap flag is set, and the database is found,
the hg will be bootstrapped. Otherwise, a new database will be created
(the new name being formed by the old name followd by an int).
Bootstrapping only works with databases containing the FULL history of
the hg.