Repare handling of InternalTransactions in node/core
BREAKING CHANGE for ProxyHandlers. Bump Babble and Dummy versions
to 0.4.3 and 0.4.2 respectively.

The Accepted field of the InternalTransactions was being ignored.
The initial thinking was that the CommitResponse would contain only
those InternalTransactions that were accepted, and we forgot to
update the node/core when this was abandoned. Also the code in
ProxyHandlers was broken because, InternalTransactions being passed
by value, the .Accept() method wasn't really doing anything. So any
application that actually looks at InternalTransactions to add or
remove peers will need to update their OnCommit ProxyHandler.