Refactor core "peers" and "validators"
The core object tracks two lists of nodes:

 - peers: the nodes that we are actively gossipping with

 - validators: the latest recorded validator-set (people who
   actually participate in consensus).

These two lists are not necessarily equal. A node which attempts to
join an existing network will start with "validators" set to the
genesis validator-set, and "peers" set to a subset of existing nodes.
As it catches up with the other nodes, with or without fast-sync,
the "validators" field tracks the hashgraph's latest recorded
validator-set.