WIP - sync-limit and fast-sync changes
We used to switch to the CatchingUp state when a SyncResponse was
over the sync-limit.

Now the use of sync-limit has changed and the state transitions as
well.

SyncLimit is used to set a max for the number of events returned in
a SyncResponse. The "responding" node computes the min of its own
SyncLimit and the requester's SyncLimit, and that's the max number
of events it will return out of the computed diff.

Once a node is part of the validator-set (upon detecting that it
belongs to the genesis.peers or after a succesfull join) it enters
the Babbling state or the CatchingUp state depending on whether
FastSync is enabled. If FastSync is enabled, it will FastForward,
and then transition to the Babbling state. Otherwise, it goes
straight to the Babbling state. The FastForward method has an extra
step to determine the best peer to fast-forward from. Indeed we
don't want to fast-forward from a node that is behind with respect
to the rest of the network.