Implemented the same CORS settings as EVM-Lite does.
Whilst it feels like potentially unnecessary clutter,
these are public endpoints . And the client side hack-arounds required
to circumvent CORS are not really production code.

I hit issues when investigating the Dashboard Code -
if we extend the evm-lite-js to cover all Babble endpoints it will
become a significant issue ongoing.