Force babble.initKey to error when no priv_key or wrong permissions
The initKey function used to automatically generate a new key when
a priv_key file was not found. We decided that it would be better
to avoid this behaviour, and expect the user to generate a key
before running Babble; like ssh. In fact we also took inspiration
from ssh to check that the key file has user-only permissions.