Do not process errors in service ListenAndServe
When Babble is used in-memory, the application could be interested
in exposing its API on the same endpoint (address:port) as Babble.
For example, this is the case in monetd. The pragmatic solution we
adopted here is to register all handlers on the DefaultServeMux,
and ignore the error when ListenAndServe is called the second time
on the same endpoint.