Replace logrus.Logger with logrus.Entry
We wnat log entries from child objects to inherit fields from their
parents. The only way to do this with logrus is to use Entries
rather than Loggers.

We also introduce a new logrus formatter to handle a prefix in
entries. The idea is to use the prefix "Babble" everywhere. This
will help to disassociate Babble logs from other logs (ex evm-lite
and monetd).