Add condition to record heads
Call RecordHeads if Seq < 0. This forces a joining node to record
heads upon joining. We were encountering a situation where a joining
node took long enough to catch-up that the other nodes had time to
commit its TargetRound (the round where the joining node gets added
to the validator-set). So the joining node was not recording heads
immediately which opened the possibility of creating a self-event
with no other-parent, which would then have round 0, which breaks
when the inmem-cache is too small.