Roll back changes to EventBody marshalling function
The recent changes are not backwards compatible. Encode and Marshal
of the golang json packages do not return the exact same bytes
(Encode adds a new line). So we were getting different hashes and
signatures for Events that were otherwise identical. The whole
hashing/signing setup should be refactored, but here we just roll
back the change to ensure our next release is compatible with the
previous one.