Reabilitate database writing in Store ParticipantEvents methods.
We need to attempt reading from the database when the inmem-store
returns an error. When a peer attemps to join a network that
already has more Events than can fit in the inmem caches, the Diff
(which determines the events that need to be returned as part of a
sync request) would not be able to retrieve "earlier" events and
the joining node would never be able to catch-up.