Use advertise address in JoinRequest
JoinRequests were using transport.LocalAddr as the joining peer's
address. This was always an IP address (public or private). We
wanted to be able to pass hostnames as well. So we added an
AdvertiseAddr to the net.Transport interface to expose an advertise
address that can be a hostname. This is used in the JoinRequest.