Experimenting with WAMP
WAMP (Web Application Messaging Protocol) is a WebSocket subprotocol
that we can use in the WebRTC signaling system. There is a server in
the middle that relays RPC requests between peers. We could implement
an offer/answer mechanism this way. net/wamp contains a minimal test
where one peer calls a method on another peer via a wamp server.